Groundwork for realtime signal display
This is horrible, but I cannot come up with a better idea. Because lvgl
insists that it must keep the whole display under control, it's
impractival to use it to display realtime data in small segments, as it
arrives. So, at each step, we first call lvgl's update function, and
after that (still under the mutex), draw a segment directly to the SPI
driver, totally sidestepping lvgl. This may end up in a mess. So we will
try to make sure that lvgl never touches the part of the screen that is
updated by the direct access.