Do not rely on GPIO0 interrupt from the button
It looks like in this board (both lilygo and waveshare), if it was at
least once used with wifi enabled, gpio0 starts to produce spurious
interrupts (possibly because the pin is shared with SPI flash), that
results in immediate shotdown.

SPI sharing does not affect reading GPIO level, so instead of using
an interrupt handler, poll interrupt level in the main loop, and act
once measured level is zero.

This will allow the app to run on boards that were previously used
with some other, wifi-enabled, app.